+++ e2e-tests/ui-components/vanilla-test-server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # e2e-tests-ui-components-vanilla-server
 
+## 0.0.28
+
+### Patch Changes
+
+- @evervault/browser@2.37.1
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [4928702]
+- Updated dependencies [cd40338]
+  - @evervault/browser@2.37.0
+
 ## 0.0.26
 
 ### Patch Changes