+++ examples/collect-card-details/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-ui-components
 
+## 0.2.3
+
+### Patch Changes
+
+- @evervault/browser@2.37.1
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [4928702]
+- Updated dependencies [cd40338]
+  - @evervault/browser@2.37.0
+
 ## 0.2.1
 
 ### Patch Changes