+++ packages/themes/CHANGELOG.md
@@ -1,5 +1,19 @@
 # themes
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [446bbcb]
+  - types@0.7.0
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [cd40338]
+  - types@0.6.0
+
 ## 0.1.5
 
 ### Patch Changes